+++ requirements.txt
@@ -5,3 +5,4 @@ psycopg2-binary
 gunicorn
 passlib
 flask-login
+flask-bootstrap
