+++ routes.py
@@ -1,0 +1,75 @@
+from flask import render_template, redirect, url_for
+from flask_login import LoginManager, login_user, current_user, login_required, logout_user
+from forms import *
+from app import app, db
+
+# Configure flask login
+login = LoginManager(app)
+login.init_app(app)
+
+@login.user_loader
+def load_user(id):
+    return User.query.get(int(id))
+
+
+@app.route("/", methods=['GET', 'POST'])
+def index():
+    login_form = LoginForm()
+
+    # Allow login if validation success
+    if login_form.validate_on_submit():
+        user_object = User.query.filter_by(username=login_form.username.data).first()
+        login_user(user_object)
+        if current_user.is_authenticated:
+            return "Logged in with flask-login!"
+
+        return render_template("index.html")
+
+    return render_template("form_simples.html", form=login_form)
+
+@app.route("/private", methods=['GET', 'POST'])
+@login_required
+def private():
+
+    # Usar isso se na for usar o "@login_required"
+    #if not current_user.is_authenticated:
+    #    return "Please login."
+    return "something"
+
+@app.route("/logout", methods=['GET'])
+def logout():
+
+    logout_user()
+    return "Logged out"
+
+@app.route('/register/', methods=['GET', 'POST'])
+def register():
+    reg_form = RegistrationForm()
+
+    # Updated database if validation success
+    if reg_form.validate_on_submit():
+        username = reg_form.username.data
+        password = reg_form.password.data
+
+        hashed_pswd = pbkdf2_sha256.hash(password)
+
+        # Add it into DB
+        user = User(username=username, password=hashed_pswd)
+        db.session.add(user)
+        db.session.commit()
+
+        return redirect(url_for('index'))
+
+    return render_template('register.html', form=reg_form)
+
+@app.route('/search/')
+def search():
+    return render_template('search.html')
+
+@app.route('/project/')
+def red_search():
+    return redirect(url_for('search'))
+
+@app.route('/project/<project_id>')
+def project(project_id):
+    return render_template('')
