+++ temp.txt
@@ -1,0 +1,21 @@
+This comes from layout.html
+
+<div id="navbar" class="navbar-collapse collapse">
+    {% if logged %}
+    <form class="navbar-form navbar-right">
+        <p>Logged in!</p>
+    </form>
+    {% else %}
+    <form action="{{url_for('index')}}", method="POST" class="navbar-form navbar-right">
+        <div class="form-group">
+        {{ login_form.username(class_='form-control', placeholder='Email', autocomplete='off', autofocus=true)}}
+        </div>
+        <div class="form-group">
+        {{ login_form.password(class_='form-control', placeholder='Password')}}
+        </div>
+        <input type="submit" value="Sign in" class="btn btn-success">
+        <a class="btn btn-primary" href="/register">Register</a>
+        {{ login_form.csrf_token }}
+    </form>
+    {% endif %}
+</div>
