+++ templates/form_simples.html
@@ -1,0 +1,16 @@
+<form action="{{url_for('index')}}", method="POST" class="form-signin navbar-form navbar-right">
+    {{ form.username(class_='form-control', placeholder='chuck@norris.com', autocomplete='off', autofocus=true)}}
+    <ul>
+        {% for error in form.username.errors %}
+            <li>{{ error }}</li>
+        {% endfor %}
+    </ul>
+    {{ form.password(class_='form-control', placeholder='Password')}}
+    <ul>
+        {% for error in form.password.errors %}
+            <li>{{ error }}</li>
+        {% endfor %}
+    </ul>
+    <button id="register" class="btn btn-lg btn-primary btn-block" type="submit">Register</button>
+    {{ form.csrf_token }}
+</form>
